requirements.txt
google-adk>=1.18.0
google-generativeai>=0.8.0
python-dotenv>=1.0.0
fastapi>=0.104.0
uvicorn>=0.23.2
python-multipart>=0.0.6
PyPDF2>=3.0.0
python-docx>=0.8.11
pandas>=2.0.0
